test: 테스트용 TestDeck 클래스 분리, Deck 사용 테스트에 적용
